requirements.txt
jdcal==1.0
lorem-ipsum-generator==0.3
mutagen==1.27
MySQL-python==1.2.5
numpy==1.9.1
openpyxl==2.1.4
pyalsaaudio==0.7
pydub==0.9.5
pyechonest==9.0.0
requests==2.5.1
simplejson==3.6.5
six==1.9.0
subprocess32==3.2.6
transliterate==1.7.3
web.py==0.37
